+++ client/src/components/buttons/GithubLoginButton.tsx
@@ -8,7 +8,7 @@ export default function GithubLoginButton() {
 
   return (
     <button
-      className="hover:bg-gray-600 my-2 flex items-center gap-2 rounded-lg bg-aod_black px-4 py-2.5 text-sm text-aod_text"
+      className="my-2 flex items-center gap-2 rounded-lg bg-aod_black px-4 py-2.5 text-sm text-aod_text hover:bg-opacity-50"
       onClick={login}>
       <FaGithub size="1rem" />
       Login with GitHub
